refactor: Move utilities to its own folder
