--- frontend/src/lib/index.ts
@@ -1,3 +1,0 @@
-export { default as Modal } from "./ui/modal.svelte";
-export { default as JsonCodeBlock } from "./components/json-code-block.svelte";
-export { default as UiTooltip } from "./ui/tooltip.svelte";
